LPD-58307 Move to the exception.handler package
