+++ modules/apps/export-import/export-import-report-service/src/main/java/com/liferay/exportimport/report/internal/exception/handler/ImportStagedModelExceptionHandlerImpl.java
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.exportimport.report.service.impl.handler;
+package com.liferay.exportimport.report.internal.exception.handler;
 
-import com.liferay.exportimport.kernel.exception.ImportStagedModelExceptionHandler;
+import com.liferay.exportimport.kernel.exception.handler.ImportStagedModelExceptionHandler;
 import com.liferay.exportimport.kernel.lar.ExportImportClassedModelUtil;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.kernel.lar.PortletDataException;
