+++ modules/apps/export-import/export-import-report-test/src/testIntegration/java/com/liferay/exportimport/report/internal/exception/handler/test/ImportStagedModelExceptionHandlerTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.exportimport.report.service.test;
+package com.liferay.exportimport.report.internal.exception.handler.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.expando.kernel.model.ExpandoBridge;
