+++ portal-kernel/src/com/liferay/exportimport/kernel/exception/handler/ImportStagedModelExceptionHandler.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.exportimport.kernel.exception;
+package com.liferay.exportimport.kernel.exception.handler;
 
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.kernel.lar.PortletDataException;
